refactor: 서비스명 변경 (zzaekkac -> jjakkak)
