+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'schedule'
+rootProject.name = 'jjakkak'
