+++ src/main/java/com/dnd/jjakkak/JjakkakApplication.java
@@ -1,0 +1,13 @@
+package com.dnd.jjakkak;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class JjakkakApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(JjakkakApplication.class, args);
+    }
+
+}
