--- src/main/java/com/dnd/zzaekkac/ZzaekkacApplication.java
@@ -1,13 +1,0 @@
-package com.dnd.zzaekkac;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ZzaekkacApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ZzaekkacApplication.class, args);
-    }
-
-}
